imported all reducers as * into indexjs file
